refactor(budget): 予算・執行セクションを共有コンポーネント化し両ページで採用
再委託の予算・執行展開表示がメインと別物（平坦リスト vs 会計集計＋カード）
だったのを解消。メインの予算・執行アコーディオンを
client/components/BudgetExecutionSection.tsx に抽出し、両ページで採用。

- 会計区分集計（一般/特別バッジ＋金額）・歳出項目カード（会計/勘定/項/目メタ
  グリッド・補足）・2-1/2-2差分警告・ドラッグ高さ変更を共有
- getAccountBadgeStyle を app/lib/account-badge.ts に共有化
- 再委託の budgetSummary 型を完全な BudgetSummary に（accountSummaries を利用）
- ドラッグ高さ変更はメインのみ（onResizeStart 指定時）。再委託は固定高さ
- 両ページで予算・執行の展開表示が同一になることを Playwright で確認
- lint/tsc グリーン、ページエラーなし
